refactor: context API 관련 커스텀 훅 경로 변경 및 적용
