+++ src/pages/Detail.tsx
@@ -11,7 +11,7 @@ import Header from '../components/header/Header';
 import { Helmet } from 'react-helmet';
 import { editPostingTime } from '../shared/sharedFn';
 import { postingsAPI } from '../shared/api';
-import useUserState from '../hooks/useUserState';
+import useUserState from '../hooks/contexts/useUserState';
 import { TypeComment } from '../typings';
 import usePostDetailQuery from '../hooks/queries/usePostDetailQuery';
 import usePostLikeHandler from '../hooks/queries/usePostLikeMutation';
